Let CMake work with subprojects.
This is a very usual use case and a place where many people found "buggy" the way to work because when importing a subproject made KDevelop to mess up 
because of the non-parsed CMakeLists.txt files placed in parent directories. Now KDevelop looks if there are and it parses them too.
**Warning** it will always look for it, it is not yet configurable in the project file (it is configurable in the .kdev4/project.kdev4 file, should be 
moved)

Now it is possible to parse a cmake subproject without parsing the other subdirectories below or in the same level. It can be useful when you are 
working on a large project and you don't need the whole to be supported by KDevelop (for instance, when I'm working on KAlgebra in KDE-Edu, I don't 
need, let's say, the Marble buildtool information).

CCMAIL: kdevelop-devel@kdevelop.org