Expose the deeper reason behind failure of Veritas::StubCreationTest::nonVirtualMethod() with some additional test commands in TestDUChain.
Addition of a single '\n' can swing the virtual-ness of a memberfunction (with no 'virtual' keyword present).
This used to run until after gsoc, so it's something pretty recent. Unfortunatly at the point it started failing there was no continuous test-report yet, so I can't easily pinpoint the commit that introduced this bug.

@note the manual test in languages/cpp/veritas/examples/stubgeneration.h is currently broken because of this, too.