- Allow statically overriding the TopDUContext-features needed for a specific url, in ParseJob. - Add a central function to ParsingEnvironment that determines whether the context needs to be updated due to not satisfied features of an import.
Enable intelligent use-searching that includes updating the duchain when required:
In the uses-widget: Collect all potential places where the use of a declaration can happen using the duchain include-structure, compute all files that need to be built/updated with complete use-information, and query an update of those files.
Once the update is ready, add the uses to the widget.

This only requires the simple processing enabled by "Full project parsing" to enable quite reliable use-computation across all loaded projects now. This btw. has nothing in common with a simple "grep", but is based on semantic language understanding.

Warning: This is quite slow when you use it on a very central declaration like QDialog..