Improve usability of the new project wizard. Automatically append the project name to the url, until the user edits the url. From that point on it accepts the users choice and doesn't change the url anymore.
I'm not sure wether this is the best way of doing it or not, so CC'ing
the list to get some more input/feedback.

Another option I currently see is allowing to edit the url, but going
back to appending the appname once the user again changes the
application name. Is that more intuitive?

CCMAIL: kdevelop-devel@kdevelop.org