Support running programs as other users, using kdesudo.
Reads the entry "Run As User" in the project config file, which currently has no gui.

For the bugreport, the only thing missing (apart from the gui to configure) is to make gdb run the process as the given user - maybe gdb itself has to be run as that user?

CCBUG:60165