Friendliness++
Add a vertical spacer to the bottom of the project info page. Looks better
that way, IMHO.

Since we're using the build manager plugin's name in the open project box,
make them more human friendly so instead of "KDevelop CMake Manager", you
see, "CMake"

Change "Manager" in the UI to "Build System". The generic importer is used
for the 'None' build system. I couldn't figure out a way to nicely deal with
the Kross plugin, so I just left it as is.


svn path=/trunk/KDE/kdevelop/plugins/managers/qmake/; revision=901581