Fixed plural form usage.  In English it's safe to use:
("%1 items", n)

...as we know that if n >= 2, there's only one plural form.  In other languages, however, we hit two problems:

(i) different plural forms depending on the value of n
(ii) the singular form is used if n=21, 31 etc.

Hence, the plural call is necessary.  I've also added a little context, as not every translator will be familiar with 
the use of the word 'overload' as in C++.

Keep up the great work you're doing on kdevelop - I've lost track of how many hours of work it has saved me :).