Refactor the Generic Project Manager to support remote projects via KIO::listRecursive
TODO: dfaure confirmed my assumption that we have to fork KIO::listRecursive to support filtering efficiently:
currently the list job traverses into filtered directories and reports useless paths with potentially many files, since it has no idea of what we are filtering.

TODO: on project reload the items directly in the project folder loose their style / icons - Apol, any idea what's happening there?

thanks to dfaure and apol for their numerous help

thanks to adymo for his hospitality during the Hack Sprint in Ukraine - it was a blast and I hope to repeat this experience as soon as possible!