Allow to disable ownership of the created widget in Sublime::View. This is apparently needed for some KParts to work properly as they expect to keep the onwership of their widgets (and react very badly if the widget is deleted before the kpart itself).
However Kate's document/view setup doesn't seem to have such a problem
and here we probably need to delete the widgets ourselves. So take
ownership for Text documents, but not for Part documents.

Note: I didn't valgrind this, so it might create a leak with some
kparts, but in that case I think the error is in the kpart which should
take ownership of its widget.
BUG:194035