Revert "Step towards 'useful' multiple-mainwindow support: Add an interface for plugins to create per-mainwindow actions, and manage+merge those actions within the additional mainwindow." and related commits.
Its quite apparent that multi-mainwindow support isn't quite that easy
to implement, so this should be done in a branch until its working -
including conversion of all plugins.

Currently KDevelop either crashes on startup (or when creating the
second window) or its missing the actions from the converted plugins.
Additionally the API discussions are not finished yet as far as I could
see.

CCMAIL:kdevelop-devel@kdevelop.org