Refactor stack frame getting.
The primary reason for this change is that with current GDB,
it is no longer required to actively ask for the list of threads.
Insteads, one can listen to the notifications that are emitted
whenever a thread is created, or exited. The current interface,
specifically IStackController::fetchThread, on the other hand,
assumes active polling. It seems more reasonable to move all
responsibility to FrameStackModel, so that debugger-specific
children classes of it can either pull, or use notification,
or other methods.

Also, FrameStackModel was associated with IDebugController.
It makes more sense to associate it with a session, so that
when we eventually support debugging two programs at the same
time, each can have it's own thread.

So, this patch kills IStackController, moving responsibility
to FrameStackModel. GDB is not yet switched to use notifications,
this is a topic for the next patch.