Create watch expressions in the context of the selected thread.
Now, if you select a thread in framestack widget, and add
variable, it will be evaluated in the thread you have just
selected. Unfortunately, if you hit a breakpoint the thread
that has hit the breakpoint is not auto-selected yet. Also,
you can select a frame as you like, and that will have no
effect.