Make thread switching work.
The primary goal of this patch is to make the "Watches"
item in the variables view update when a different
thread is selected by the user. 

This also does lots of refactoring in update logic for
frame and variables, switching it to use event as opposed
to states. The reason is that change in current thread and
frame does not result in any change of IDebugSession::state,
and therefore we can't detect this.

CCBUG: 207167