add Import kind to Declaration, so we can re-use all the nice Declaration features for file includes, e.g. uses, jump to dec==file etc.
To use, just add the declaration to the _included_ file, I personally add it directly to the beginning, with an empty range.
Than just build uses for every include. Works very well.

TODO: port the code to create a nice navigation widget from the IncludeItem to KDevplatform, so others can re-use it.