Disable the sourceformatter as its not possible to fix this up without major time investment.
I've looked a bit at the code today and all I can say is wow,
this is broken. It stores pointers to plugins, doesn't update them as needed
assumes everywhere that things will be alright (i.e. no empty-checks, no
null-pointer-checks). This is not possible to fix easily without changing
major parts of the code. At this point in the release such a major
restructuring is not going to happen, additionally this codebase is basically
unmaintained ever since it was imported into svn.

So we'll have to live without it for 4.0