Revert removal of s_appRunning.
This change broke adding breakpoints while application is running,
and had other bugs. The ideal solution is to add support for non-stop
mode, including keeping per-thread 'running' state and automatically
picking a suitable thread to use for inserting breakpoint. However,
such changes are too big, and we're in feature freeze.