Correctly focus view's widget when switching views using the tabbar. This fixes following problematic scenario: - switch to one of the views by clicking on the tabbar - execute any menu action which shows a dialog (File->Open, Help->About, etc.) - watch current tab lose focus
The problem was that we shouldn't have called setFocus() for a widget before activating its view.
Together with other focusing mechanisms, this made Qt remember that the previously focused widget
in current window was not the newly switched tab, but the previous one.

In the scenario above, when application closed the popup, it returned focus to the previously
focused widget in the mainwindow. Previous tab's widget was considered as previouly focused
and it got focus instead of the current tab.

BUG:190748