Add "add" and "remove" buttons
editable comboboxes don't work well within kcm's as <Return> will trigger
the "Ok" button and hence not be handled by the combobox (because its
implemented via a shortcut). So need a way to add the text in the combobox
lineedit to the list of items and of course a way to remove one too.