Fix storage of current config for new configs.
When creating a new config it initially has no grpname. The name is
generated only once the storing to disk happens. At that point we
need to properly update the original config in the list with the one
having the fixed grpname. Additionally the grpname needs to be put
into the combobox data and then we can safely write out the current
config to disk.