Move the lessThan implementation from proxy to project model.
The reason is that custom item subclasses may want to define their
position in the tree themselves. This is only possible if they can
decide themselves which other items are "greater".

Additionally do not default to "File" type in baseType, instead use the
real type. This fixes crashes with custom item subclasses (file() was
called on them without checking the resulting pointer).