Remove the proxy-selection model and the public project-selection-model.
This causes problems for dynamic updates of the model, which can happen
when a top-item is added immediately and items are added later on. Also
some functions of the itemselection model are not virtual, which probably
would need to be virtual to really create a proxy-selection-model.

Nothing uses the public project-selection-model so far, so shouldn't be
a problem.