Fix project view crashing on startup.
For some reason the project treeview crashes when calling the setResizeMode
on the column 0. The assert suggests that the header doesn't know about
the column yet. Not sure whats going on, but as there's a single column
only lets activate resize-to-contents for all.

Additionally removes some uneeded code and moves the setStretchLastSection
to the ui file.