Don't let -stack-list-locals get all values as we don't use them.
Instead we use -var-create on every variable to get the values. It would
however make sense to use the values from -stack-list-locals for better performance.

As a side effect this fixes problems where -stack-list-locals failed with an error
that a pretty printer of a variable in the stack caused; like the uninitialized std::map
in the referenced bugreport.
BUG: 229862