clear strings when looking for identifier under cursor, but only after we checked for an #include
this prevents the Macro-Tooltips been shown when one has something like
 #define FOO;
 const char* s = "FOO";
We might want to clear comments as well, but since we only operate on a single line (performance!),
it won't work for multiline comments. But imo it's not a big deal considering one might document
some feature and put the Macro into the comment.