Fix closing of changed documents.
When closing documents which had unsaved changes, kdevelop would close
it before it showed the question-dialog to save/close or cancel.

It introduces a new problem, that is it asks for confirmation more than
once, but since the bug fixed by this can loose your data I'm comitting
it none-theless for the 4.0 release.

BUG:228774