Fix compile warnings related to deprecated hash_map
g++ >= 4.3 has the unordered_map header/class as replacement for the now
deprecated hash_map header. This commit takes that in account to fix a
bunch of compile warnings related to this. I made sure this still works
for g++ < 4.3 and did *not* change anything related to MSVC builds. The
latter might (still) have a similar warning.

When LTS for g++ < 4.3 has ended, the include should be changed to
<unordered_map> and the equally named type should be used directly.

NOTE: The hash_map was profiled by David Noden to be better performing
than QHash. I didn't do additional profiling for unordered_map but
assume it performs equally as hash_map.