Add unit tests for cmake_minimum_required AST creation.
It makes sure the AST contents are actually filled properly, instead of
only checking if it accepts valid code and rejects invalid code. Also added
a test for a three-component version number.

This test exposes an actual bug in the AST creation, so I marked a QCOMPARE
as expected fail.