Get rid of all static and globally allocated objects.
This patch removes all static objects and all other stack-allocated
objects created when one of the compilation units in language was loaded.

The reason for this is so that we can get rid of the second binary, we
are able to move the duchain data into the session directory and we have
better-defined construction order which will help with moving to other
platforms.