properly reset function parameters in ExpressionVisitor::visitFunctionCall
Before a call to a default ctor in a function call would mess up the parameters
of the outer function call.

Add test which in itself shows interesting behavior, David please take a look:
When the patch (that looks perfectly fine) is run alone it works like a charm.
When run together with the other tests in the DUChain Test Suite it fails. If
one compares the output of the AST Dump one sees that it's completely different
in the two cases... A severe bug somewhere?

Thanks to Tobias König for the patch

CCMAIL: david.nolden.kdevelop@art-master.de