Drop QStandardItem* dependency.
This dependency is bad as it makes the project model require gui code to
be linked in. This patch replicates the QStandardItem* API as far as
existing code in KDevelop/KDevPlatform needs it manually. Also the
unecessary extra Private-Classes are dropped and instead the url is handled
in the base-class.

This change also disables kross-wrappers for now until I can regenerate
them and the code that edits items until its ported.