Add a small gui app to test the model performance.
This is a very minimal gui that shows the model data in a tree and allows
to add items to the model and expand/collapse it. The idea of this utility
test is to measure the performance of these operations to find out which
parts of the model/items are slow (in particular when doing significant
changes to the implementation).