Verify that the project name is used for the topitem text.
This verifies the fix I've just added to restore old behaviour. The top
item is expected to use the project name not the foldername so the display
in the tree shows the project info as the foldername can be very easily
conflict with another projects foldername (trunk vs. trunk from two
svn-checkouts). This broke quite some code so a unit-test is necessary.