Add a small test for creating items in a background thread.
This merely verifies that our invokeMethod code works as we want it to.
It also already shows a problem with the existing code, due to using
AutoConnection Qt chooses a QueuedConnection when creating items in a
background thread. This breaks at least the model test because it expects
to still get the original model data when rowsAboutToBeInserted is emitted,
however in our example here all the changes have already been done when
the signal is emitted.