Implement support for re-parsing project files after changes.
I'm not sure this works really as its expected to be (i.e. I'm not sure
it fully reparses the sources), but that would be a bug in the language
library in kdevplatform then.

Now upon switching the builddir or doing other changes a complete reparse
of the project is started (if the option to parse all sources is active),
so afterwards all parsed information should be up-to-date.