Code cleanup.
After checkout it's up to the project manager to decide if it has to be reloaded or what. I think it is good that VCS notifies this kind of stuff but we have to do it better and it was not being used anyway so we'll get there. Also implementation was wrong (was closing and opening the project instead of using ::reloadModel).