"Apply" button enabled (2 pass replacement) - Root item added
The root item allows a fast check/uncheck of all matches. The updateCheckState
logic has been generalized so now, it can work on any level of nesting (if
we decide to reproduce directory trees for instance)

Also, on a simple search, "Apply" button is disabled and checkboxes are no longer
displayed.

When an item has been replaced, it is disabled, but user can check other items and
apply again.