Improve version check for KDevPlatform
This enhances the version-checking by marking different minor versions
of KDevPlatform as incompatible with one-another, i.e. 1.1 is
incompatible with 1.2 (1.1.60 and later). So a project that requires
kdevplatform 1.1 will not be buildable against 1.1.60 and later versions
of KDevPlatform.

Ideally we could do this inside a custom-written
KDevPlatformConfigVersion.cmake.in, but if that is done cmake produces
the unhelpful "Could not find configuration file or package for
KDevPlatform" error which doesn't say anything about this problem.