Fix a crash when loading a project, by initializing static data
This prevents a crash I can reproduce pretty reliably with a closed project
and the custom-buildsystem plugin. Every second startup KDevelop crashed
due to two threads trying to access the recursiveimportrepository, which
caused one thread to get a null-pointer in its QFile member. Initializing
the repository during startup fixes that for me.

BUG: 250779