Fix updating of status icon when switching views.
I'm not sure wether this was working correctly when using the tabbar
but it was broke without using that. The label got the wrong icon all
the time as it was never properly updated when switching views or closing
one.

Additionally we now use the same API in all places so changes to how the
text is shown in either case are taking proper effect.