Show the complete path to documents when no tabbar is used
Without a tabbar there's enough space for the complete url/path to the
file currently open. So lets add that as it makes it easier to identify
the file one is looking at. Currently uses absolute path/url as sublime
has no access to kdevplatform API, needs some more changes to do that.

BUG: 259132