Make the filename corner variable size instead of fixed.
This allows arbitrary splitting of views without increasing the mainwindow
size because the filename label can make itself smaller if necessary. This
is achieved by eliding the text through KSqueezedTextLabel.

Note that its necessary to change the members declaration accordingly so
that the overloaded setText function of KSqueezedTextLabel is called,
otherwise the label will stay empty.

BUG: 259909