reintroduce the cleanup of 986ec7bb18211a7789e28c48fdc750668ae9b1d5 (no addViewSilently, no window->setArea in working set loadToArea), but fix the problem described in 2f586e506ac59153b0b95e008acc88cea490e7d3 (restoring of split views)
the problem was that the working set adds views and the sublime code up to now expected indizes being
reconstructed in a top-down manner (i.e. first the parent index, then its children, ...)

now if there isn't a splitter found for a parent index, we just do that first. it seems to work properly in my tests.