ProjectManagerView state save/restore
Added saving the state ot the splitter controlling sizes of the
ProjectTreeView and ProjectBuildSet into a session.

Added default strach factors for this splitter items: 75% for the
ProjectTreeView and 25% for the ProjectBuildSet.

Theoretically this is not the correct fix but rather a workaround
that just happens to work for most users.

REVIEW: 100584