Fix standard include dirs and macros under Windows This patch enables Kdevelop to find the include directories and standard macros under Windows.
There is no "gcc -dM -E -" under Windows. So you need to do a hack to find the standard defined macros. Therefore a empty cpp file is compiled with a special commandline option "/Bx" to replace the compiler against a program which interpretes the output of the preprocessor.

For include directories Windows uses enviroment variables, so this patch processes them.

This patch has been written by Patrick Spenderin, thank you for that.