Move Diff2::Marker::operator== from a unit test to the class itself.
The overloaded operator in the unit test file was being declared
in the global namespace, which in theory means it won't be found
by C++ name-lookup rules. gcc incorrectly accepts the code anyway,
but clang doesn't, so this caused compiler errors.

Either way, it seems nicer to have the operator in the class itself,
so that it's available to all users of the class.