add isExplicitlyDeleted to Declaration
This is useful to mark declarations as explicitly deleted,
as possible in C++2011 with the deleted function definitions.

Other languages might have something similar, e.g. to hide
inherited declarations.

As this increases the number of size-one bitfields in DeclarationData
to 8, the disk-format should stay compatible.

Also display the flag in the details section of the navigation context.