Revert "enable final on the cpp subtree."
this patch contains unrelated changes which can cause
crashes, esp. the failed rewrite of foreach -> for is utterly wrong

please put it on reviewboard first for a proper discussion
on why this is required. if I understand it properly, only
the namespace change is actually required to enable final compilation?

This reverts commit 6035cf16494f59e73ec423377b0ea590c69387d5.