Fix cleanup of cache-model
Directly deleting the cache model and leaving a reference in the view
is dangerous as the related bugreport shows. I'm not sure why there was
an event sent out to the selectionmodel, but it seems this happens already
when deleting the model. The event however is delivered later on in the
nested event-loop from the kmessagebox and at that point the selectionmodel
is already gone.

So disconnect the selectionmodel properly, explicitly remove the model
from the view  and let the cache model alive for one round of
event-processing.

CCBUG: 261519