Add todo about code I don't understand.
This extra step through the findProjectForUrl shouldn't be necessary since
the module already knows the project it was created for. I think this
is just from some old conversion from multiple dialogs to a single one,
but I don't dare to touch this without some further clarification from
the author. So just add a todo.