Fix creation of builddirs with custom cmake args
This should fix all open problems configuring new builddirs and changing
stuff in the builddir-creation dialog. It takes the values from the dialog
and explicitly initializes the kconfigskeleton items from them. This
ensures the values end up in the config file and we don't loose things
set in the dialog.

Also necessary for this is to pre-initialize the local variables in save
so that there's no empty stuff written out in case no model has been
created yet.

Last add a TODO about the possibility to have different builddirs with
different configs and not being able to re-run cmake from scratch for
them, but thats something for another time. Probably would be useful to
completely drop the kconfig-skeleton stuff since its not used anyway
by the kcm and would make it more explicit that the module has to keep
track of these things manually.

BUG: 261519