Fix possible crash on shutdown
In my case the mainwindow was deleted at a point where either the active
session or the the core was already gone (the backtrace did not include
sufficient information, but it crashed in ~GrepOutputView) or access to
kconfig was maybe not possible anymore.

So de-register the toolview during plugin unloading which triggers removal
of the toolview. This is early enough so that the configuration stuff
can happen without crashing.