c++11 support: folow spec and use/introduce expression-list where required
this way we can finally pass init-lists as arguments to a function call

NOTE: binary-expression is not (anymore?) in the spec and should probably
be reomved. Also the m_arguments* code, esp. related to binary-expressions
in the expressionvisitor should be cleaned up now as it's most probably
obsoleted by the new (simpler) code.

NOTE2: also fixes an issue with variadic template pack-expansion in
function pointers