Don't require cmake-folders for source-files.
Its perfectly valid to have source code in non-cmake folders, so lets
handle the case of not being able to find a cmake-folder when searching
for includes/defines gracefully. Otherwise the assert triggers shutdown
of Kdevelop when having a project that has a directory with no
CMakeLists.txt but plenty of source code files.