Fix providing defines for cmake-projects
This fixes parsing of c++ files in subdirectories of a CMake project
when the c++ file uses defines set in a parent directories CMakeLists.txt.

i.e. when
project/CMakeLists.txt has add_definitions(-DHAVE_TEST)
project/src/foo.cpp has a #ifdef HAVE_TEST section
the C++ plugin wouldn't parse inside that section since the cmake manager
did not provide the HAVE_TEST define for foo.cpp since its set in a parent
directory.

Using the formerParent API should ensure that cases such as
add_subdirectory(a/b) work correctly, similarly to the include-dirs
calculation.

As the comment says, this is not producing 100% correct results (compared
to the actual defines when compiling the files of a target), but it guess
the result is good enough and for getting it perfect a redesign of the
codebase would be needed.

BUG: 287462