add KDevelop::TestFile
this is a helper class we used in PHP for some time already.
makes it easy to use test language plugins with the full
kdevplatform api, i.e. including full language plugin and
background parser etc. pp.

also makes it easy to test cases that require projects
and multiple files