Fix storing/restoring of environment settings
This works around a change in kdelibs that was done to fix
KConfigDialogManager's behaviour with Qt4.8 and normal Q/KComboBox. With
that change the user-property is not considered anymore for Q/KComboBox
and instead the hardcoded index/text-handling part is done. That breaks
our own KComboBox subclass since we relied on the fact that our own
user-property would override that hardcoded stuff.

After looking at the kconfigdialogmanager sources I noticed that it
supports a special property "kcfg_property" which can point to a
property-name and then is considered first as the property to store
into the config file. With that we can again restore the behaviour of
storing the currentProfile to the config file.

This will work on all kdelibs back to sometime 2009, which covers KDE
4.5 and later at least, which is enough since the code requires 4.6 anyway.