Fix build with KDE4_ENABLE_FINAL.
Turns out that if you use Q_DECLARE_OPERATORS_FOR_FLAGS inside a namespace,
it screws up the QFlags operators for anything that follows.

It was only happening with unrelated files: a header calling the macro
inside a namespace, and a .cpp that *doesn't* include that header
being affected. So the error only appeared in FINAL builds
where all .cpp files are built together.