backport from master: disable unit tests which require user interpretation
both, TestDUChain::testImportStructure and ::testImportCache
merely output statistics to the CLI. Neither checks the end-state
for validity and as such these tests are useless to be run in an
automated fashion.

Furthermore is TestDUChain::testImportCache apparently dependent
on a non-empty DUChain cache which is not the case for the unit
tests and as such even more useless.

If someone could explain to me what these tests should verify
and test I could extend and re-add them. For now, I've disabled
them.

cherry-pick from commit 7b1315ae3550dde8bb2039577c7ed7a9719c55e3