Fix build of CMake tests if source dir has spaces.
The build system was using add_definition() to pass compile-time paths
to the test code, but this breaks if the paths have special characters
such as spaces, and the needed quoting/escaping is tricky to get right
in a portable way. I changed it to use a header file that gets
configure_file'd. Strings in the C preprocessor have a more portable
syntax than shells and buildsystems in multiple OSs :)