drastically improve performance of ProjectFileItem::setUrl
KMimeType::findByUrl was actually not the main culprit, instead it was
iconName with the url passed. This is then checked to see if the url
points to a directory and then it is checked whether that directory
contains a .directory file defining an icon.

Since we are looking at a Project*File*Item, this is of course nothing
we need to do. Just ommitting the URL speeds this procedure up by roughly
one order of magnitude (a big project - Qt in this case - took before
about 0.8s for the iconName call alone, now it's down to about 0.04s
these numbers are accumulated over the whole project loading process)