Disable the multiple-toolview support for the 4.3 release
Reason: Inherent bugs which need significant work to be fixed.
Example: Open two equal toolviews (for example 2x konsole),
activate one, switch area, switch back, -> Both are
active instead of one. The problem is that views are
identified purely by their factory-id, which is equal
for toolviews of the same type.
The resulting state with 2 open toolviews looks buggy
and is unexpected and difficult to fix for the user,
this shouldn't happen in a quality release.

CCMAIL: kdevelop-devel@barney.cs.uni-potsdam.de