delay calls to QuickOpenModel::restart
this solves a crash that could happen if one accidentally opened
quickopen before a language plugin (that is also a provider) was
loaded.

this would end up in ProjectItemDataProvider::data's call to
standardContextForUrl which would try to load the language plugin

the language plugin in turn would register itself, calling
restart again, closing the loop.