skip files that are larger than 5MiB during parsing
our memory consumption is quite considerable for
large files, in the bug report below e.g. it is
roughly 30x that of the parsed files size.

to prevent asserts / OOM exceptions we just skip
files that are that large. most often enough,
these files are uninmportant generated files
anyways and don't need to be parsed to get useful
developer features. Examples are e.g. Qt resource
files.

note: not translated to get it in for 1.3. Since
this error is not very user visible / important,
I think it's ok this way. I'll properly mark it
as translatable for 1.4.

BUG: 291248