Redo the data engine's publishing behavior from scratch.
The original data engine's simplistic approach of removing, regathering
and republishing *all* data when a filesystem change occurs was a poor
fit for KDevelop, which is constantly hitting its session dirs/files for
reasons other than modifying the metadata we care about. Aside from
the general inefficiency of all this churn this also caused flicker in
the plasmoid as all items were removed and readded for no good reason.

This new version takes care to keep state internally and only publish
actual changes in the data.